+++ frontend/src/components/sidebar/StackList.tsx
@@ -4,8 +4,9 @@ import { useStackKeyboardShortcuts } from '@/hooks/useStackKeyboardShortcuts';
 import { CommandItem, CommandList } from '@/components/ui/command';
 import { Skeleton } from '@/components/ui/skeleton';
 import type { Label } from '@/components/label-types';
-import { StackRow, statusText, statusColor } from './StackRow';
-import type { StackRowStatus } from './StackRow';
+import { StackRow } from './StackRow';
+import { statusText, statusColor } from './stack-status-utils';
+import type { StackRowStatus } from './stack-status-utils';
 import { StackGroup } from './StackGroup';
 import { StackContextMenu } from './StackContextMenu';
 import { StackKebabMenu } from './StackKebabMenu';
